Fix CMake binary name and function signature mismatches
1. Change binary name from kangaroo256 to kangaroo in CMakeLists.txt
2. Fix function signature mismatches between headers and implementations:
   - Int.h: SetBaseN/GetBaseN now use const char* for charset
   - Kangaroo.h: ReadHeader/SaveHeader/SaveWork use uint32_t for type
   - Kangaroo.h: OpenPart uses const char* for mode
   - Timer.h: printResult/getResult use const char* for unit
3. Fix sign comparison in Timer.cpp fread call

These changes resolve compilation errors and some warnings.
Remaining: 165 compiler warnings to be fixed in next commit.

+++ main.cpp
@@ -577,7 +577,7 @@ int main(int argc, char* argv[]) {
   }
 
   if(gridSize.size() == 0) {
-    for(int i = 0; i < gpuId.size(); i++) {
+    for(size_t i = 0; i < gpuId.size(); i++) {
       gridSize.push_back(0);
       gridSize.push_back(0);
     }
